// packages/plugin-predictions/src/types.ts
export interface ApiResponse {
    data: TokenData;
    success: boolean;
    error?: string;
}
// ...
export interface TradeDataResponse {
    success: boolean;
    data?: TokenTradeData | null;
    error?: string;
}
// ...
export type ParsedInput = {
    type: "cashtag" | "eth" | "sol";
    value: string;
    chain: string;
};